Adaptado para novas versões de dependencias e para usar banco de dados MySQL

+++ src/main/resources/migrations/create_product_table.sql
@@ -2,7 +2,7 @@
 
 --changeset taiza:4
 create table product (
-    id int not null auto_increment primary key,
+    id bigint not null auto_increment primary key,
     name varchar(30),
     description varchar(80) not null
 );
